+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Add new admin reports [#1391](https://github.com/open-apparel-registry/open-apparel-registry/pull/1391)
+
 ### Changed
 
 - Consolidate contributor type options [#1395](https://github.com/open-apparel-registry/open-apparel-registry/pull/1395)